refactor(auth): consolidate backup-codes into a single server module
node:crypto / bcrypt make the whole helper server-only, so fold the former pure
core (canonicalize, generation, hashing, matching) into backup-codes.server.ts
and drop the non-.server file — matches the repo convention (csrf.server.ts,
honeypot.server.ts) and removes the risk of importing node:crypto into a client
bundle. Tests move into backup-codes.server.test.ts alongside the DB-layer tests.
